+++ CHANGELOG.md
@@ -8,6 +8,29 @@ Under-the-hood changes that do not have an impact on how end-users run our proce
 increment the patch number. The rational for versioning, and exact wording is taken from BACTpipe: DOI:
 10.5281/zenodo.1254248 and https://github.com/ctmrbio/BACTpipe)
 
+## [3.1.0] - 2019-10-03
+TL;DR:
+- QoL changes to WGS workflow
+- Simplified installation by moving all tools to a container
+
+### Added
+- Benchmarking using psutil
+- ML variant calling for WGS
+- `--singularity` option to `config case` and `config reference`
+
+### Fixed
+- Fixed a bug with boolean values in analysis.json 
+
+### Changed
+- `install.sh` simplified and will be depricated
+- Singularity container updated
+- Common somatic and germline variant callers are put in single file
+- Variant calling workflow and analysis config files merged together
+
+### Removed
+- `balsamic install` is removed
+-  Conda environments for py36 and py27 are removed
+
 ## [3.0.1] - 2019-09-11
 
 ### Fixed